move config to agent
