+++ packages/adapters/watcher/src/adapter.ts
@@ -1,9 +1,8 @@
 import { jsonifyError, RequestContext } from "@connext/nxtp-utils";
 
 import { alertViaDiscord, alertViaPagerDuty, alertViaSms, alertViaTelegram } from "./alert";
-import { getConfig } from "./config";
 import { Pauser } from "./pause";
-import { Verifier, VerifierContext, AssetInfo, Report, PauseResponse, VerifyResponse } from "./types";
+import { Verifier, VerifierContext, AssetInfo, Report, PauseResponse, VerifyResponse, WatcherConfig } from "./types";
 import { AssetVerifier } from "./verifiers";
 
 // Aggregation class for interfacing with all adapter functionality.
@@ -35,11 +34,9 @@ export class WatcherAdapter {
     return await this.pauser.pause(reason, domains);
   }
 
-  public async alert(report: Report): Promise<void> {
+  public async alert(report: Report, config: WatcherConfig): Promise<void> {
     const { requestContext, methodContext, logger } = report;
     logger.info("alert: Attempt to alert", requestContext, methodContext, report);
-
-    const config = await getConfig();
 
     const errors = [];
     // attempt to alert via discord
